Fix popover on sidebar not appearing properly AB#10416
The initial state of the sidebar being open or collapsed depends on whether the app is in mobile or desktop view.  Instead of managing this behaviour inside the sidebar component, it is now handled by the navbar store.  Additionally, the app has been updated to delay rendering until after it has determined whether it is in mobile or desktop view.  Combined, these changes allow the sidebar to render with the appropriate width initially, enabling the popover to position itself correctly.